Support createPortal in core
